Change IntegerField to FloatField for Max!

+++ app/main/forms.py
@@ -1,7 +1,7 @@
 from flask.ext.wtf import Form
 from flask.ext.wtf.file import file_allowed
 from flask.ext.pagedown.fields import PageDownField
-from wtforms import StringField, TextAreaField, BooleanField, SubmitField, IntegerField, DateField
+from wtforms import StringField, TextAreaField, BooleanField, SubmitField, IntegerField, DateField, FloatField
 from flask.ext.wtf.file import FileField
 from wtforms.validators import Required, Optional, Length, Regexp
 from wtforms import ValidationError
@@ -81,7 +81,7 @@ class ChangeSettingsForm(Form):
     submit = SubmitField('Change settings')
 
 class ExpenseNoteForm(Form):
-    total = IntegerField('Total cost', validators=[Required()])
+    total = FloatField('Total cost', validators=[Required()])
     description = StringField('Description', validators=[Required()])
     bank_account = StringField('Bank account', validators=[Required()])
     date = DateField('Date (costs made on)', validators=[Required()], format='%d/%m/%Y')
